+++ src/widgets/QuickPost.php
@@ -10,6 +10,7 @@ namespace craft\widgets;
 use Craft;
 use craft\base\Widget;
 use craft\helpers\ArrayHelper;
+use craft\helpers\Html;
 use craft\helpers\Json;
 use craft\models\Section;
 use craft\web\assets\quickpost\QuickPostAsset;
@@ -175,12 +176,14 @@ class QuickPost extends Widget
             ]);
 
         $fieldJs = $view->clearJsBuffer(false);
-
-        $view->registerJs('new Craft.QuickPostWidget(' .
-            $this->id . ', ' .
-            Json::encode($params) . ', ' .
-            "function() {\n" . $fieldJs .
-            "\n});");
+        $jsParams = Json::encode($params);
+        $jsHtml = Json::encode($html);
+        $js = <<<JS
+new Craft.QuickPostWidget($this->id, $jsParams, () => {
+  $fieldJs
+}, $jsHtml);
+JS;
+        $view->registerJs($js);
 
         return $html;
     }