+++ src/gql/arguments/elements/Asset.php
@@ -106,4 +106,15 @@ class Asset extends ElementArguments
     {
         return [];
     }
+
+    /**
+     * @inheritdoc
+     */
+    public static function getStatusArguments(): array
+    {
+        $arguments = parent::getStatusArguments();
+        unset($arguments['status']);
+
+        return $arguments;
+    }
 }