+++ src/gql/arguments/elements/GlobalSet.php
@@ -50,4 +50,12 @@ class GlobalSet extends ElementArguments
     {
         return [];
     }
+
+    /**
+     * @inheritdoc
+     */
+    public static function getStatusArguments(): array
+    {
+        return [];
+    }
 }